+++ utils/__init__.py
@@ -1,0 +1,3 @@
+from .time import *
+from .discord import *
+from .words import *
